+++ example/lib/main.dart
@@ -9,6 +9,12 @@ import 'custom_button.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
+  FlutterBranchSdk.setPreinstallCampaign('My Campaign Name');
+  FlutterBranchSdk.setPreinstallPartner('Branch \$3p Parameter Value');
+  FlutterBranchSdk.addFacebookPartnerParameter(
+      'em', '11234e56af071e9c79927651156bd7a10bca8ac34672aba121056e2698ee7088');
+  //FlutterBranchSdk.clearPartnerParameters();
+
   runApp(const MyApp());
 }